refactored bitPattern rotation code pt3
