+++ src/com/sproutlife/model/seed/SeedSproutPattern.java
@@ -10,6 +10,7 @@ package com.sproutlife.model.seed;
 import java.awt.Point;
 
 import com.sproutlife.geometry.Rotation;
+import com.sproutlife.geometry.Rotations;
 
 public class SeedSproutPattern {       
     
@@ -36,7 +37,7 @@ public class SeedSproutPattern {
     }
     
     public Point getSproutOffset(Rotation r) {
-        return BitPattern.invRotateOffset(getSproutOffset(), getSeedPattern(), getSproutPattern(), r);
+        return Rotations.invRotateOffset(getSproutOffset(), getSeedPattern(), getSproutPattern(), r);
     }
     
     public Point getSproutCenter(Rotation r) {
